pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "pluto-ai"
version = "1.0.0"
description = "AI-powered code security vulnerability scanner"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "0xSaikat", email = "contact@hackbit.org"}
]
keywords = ["security", "vulnerability", "scanner", "code-analysis", "ai", "static-analysis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "click>=8.0.0",
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "requests>=2.28.0",
    "GitPython>=3.1.0",
    "reportlab>=4.0.0",
]

[project.scripts]
pluto = "pluto.cli:cli"

[project.urls]
Homepage = "https://hackbit.org"
Repository = "https://github.com/0xsaikat/pluto"
Issues = "https://github.com/0xsaikat/pluto/issues"